requirements.txt
fastapi==0.121.2
httpx==0.28.1
locustio==0.999
pandas==2.3.3
psycopg==3.2.12
pydantic==2.12.4
pytest==8.4.2
pytest_asyncio==1.3.0
SQLAlchemy==2.0.44
tqdm==4.67.1
uvicorn==0.38.0